proxy: separate proxy module from main
