+++ proxy/src/proxy.rs
@@ -1,0 +1,151 @@
+use async_trait::async_trait;
+use bytes::Bytes;
+use std::fmt::Error;
+use clap::Parser;
+use serde::{Deserialize, Serialize};
+use std::net::ToSocketAddrs;
+use log::{debug, error, info};
+use pingora::upstreams::peer::HttpPeer;
+use pingora::{Result};
+use pingora::http::{ResponseHeader, StatusCode, Method};
+use pingora::proxy::{ProxyHttp, Session};
+
+const UPSTREAM_HOST: &str = "localhost";
+const UPSTREAM_IP: &str = "0.0.0.0"; //"125.235.4.59"
+const UPSTREAM_PORT: u16 = 8000;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RequestBody {
+    data: String,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct ResponseBody {
+    data: String,
+}
+
+pub struct EchoProxy {
+    addr: std::net::SocketAddr,
+}
+
+pub struct MyCtx {
+    buffer: Vec<u8>,
+}
+
+impl EchoProxy {
+    pub(crate) fn new() -> Self {
+        let addr = (UPSTREAM_IP.to_owned(), UPSTREAM_PORT)
+            .to_socket_addrs()
+            .unwrap()
+            .next()
+            .unwrap();
+        EchoProxy { addr }
+    }
+
+    fn get_method(session: &Session) -> String {
+        let request_summary = session.request_summary();
+        let tmp: Vec<&str> = request_summary.split(" ").collect();
+        let method: &str = tmp.get(0).unwrap();
+        method.to_string()
+    }
+
+    async fn handle_request(session: &mut Session) -> Result<Option<ResponseBody>> {
+        // read request body
+        let mut body = Vec::new();
+        loop {
+            match session.read_request_body().await? {
+                Some(chunk) => body.extend_from_slice(&chunk),
+                None => break,
+            }
+        }
+
+        // convert to json
+        match serde_json::de::from_slice::<RequestBody>(&body) {
+            Ok(request_body) => {
+                debug!("Request body: {:?}", request_body);
+                // TODO manipulate body here
+                Ok(Some(ResponseBody { data: format!("Hello from echo server! - {}", request_body.data) }))
+            }
+            Err(err) => {
+                error!("ERROR: {err}");
+                Ok(None)
+            }
+        }
+    }
+
+    async fn set_headers(response_status: StatusCode, body_bytes: &Vec<u8>, session: &mut Session) -> Result<()> {
+        let mut header = ResponseHeader::build(response_status, None)?;
+        header.append_header("Content-Length", body_bytes.len().to_string()).unwrap();
+        // access headers below are needed to pass browser's policy
+        header.append_header("Access-Control-Allow-Origin", "*".to_string()).unwrap();
+        header.append_header("Access-Control-Allow-Methods", "POST".to_string()).unwrap();
+        header.append_header("Access-Control-Allow-Headers", "Content-Type".to_string()).unwrap();
+        header.append_header("Access-Control-Max-Age", "86400".to_string()).unwrap();
+        session.write_response_header_ref(&header).await
+    }
+}
+
+#[async_trait]
+impl ProxyHttp for EchoProxy {
+    type CTX = MyCtx;
+    fn new_ctx(&self) -> Self::CTX {
+        MyCtx { buffer: vec![] }
+    }
+
+    async fn upstream_peer(
+        &self,
+        _session: &mut Session,
+        _ctx: &mut Self::CTX,
+    ) -> Result<Box<HttpPeer>> {
+        let peer: Box<HttpPeer> = Box::new(HttpPeer::new(self.addr, false, UPSTREAM_HOST.to_owned()));
+        Ok(peer)
+    }
+
+    async fn request_filter(&self, session: &mut Session, ctx: &mut Self::CTX) -> Result<bool>
+    where
+        Self::CTX: Send + Sync,
+    {
+        let mut response_body = ResponseBody { data: String::from("") };
+        let mut response_status = StatusCode::OK;
+
+        // get request method
+        let method = EchoProxy::get_method(session);
+
+        // only POST method is allowed for now
+        if method == Method::POST.to_string() {
+            match EchoProxy::handle_request(session).await? {
+                Some(res) => {
+                    response_body = res;
+                }
+                None => {
+                    response_status = StatusCode::BAD_REQUEST;
+                }
+            }
+            // browser always sends an OPTIONS request along with POST for 'application/json' content-type
+        } else if method == Method::OPTIONS.to_string() {
+            response_status = StatusCode::NO_CONTENT;
+        } else {
+            response_status = StatusCode::METHOD_NOT_ALLOWED;
+        }
+
+        // convert json response to vec
+        let response_body_bytes = serde_json::ser::to_vec(&response_body).unwrap();
+        EchoProxy::set_headers(response_status, &response_body_bytes, session).await?;
+        session.write_response_body(Some(Bytes::from(response_body_bytes)), true).await?;
+
+        Ok(true)
+    }
+
+    async fn logging(
+        &self,
+        session: &mut Session,
+        _e: Option<&pingora::Error>,
+        ctx: &mut Self::CTX,
+    ) {
+        let response_code = session
+            .response_written()
+            .map_or(0, |resp| resp.status.as_u16());
+        // access log
+        info!("{} response code: {response_code}", self.request_summary(session, ctx));
+    }
+}
